♻️ (GEK): NewSystem - move TSSEmptyCoordinator to preview file
add #if DEBUG guards to struct + previews
