+++ Modules/GameEngineKit/Sources/_NewSystem/Views/TTS/TTS/Components/1_TTSOneChoice.swift
@@ -6,6 +6,8 @@ import Combine
 import ContentKit
 import SwiftUI
 
+// MARK: - TTSView.OneChoiceView
+
 extension TTSView {
     struct OneChoiceView: View {
         @ObservedObject var viewModel: TTSViewViewModel
@@ -22,27 +24,11 @@ extension TTSView {
     }
 }
 
-#Preview {
-    // MARK: - TTSEmptyCoordinator
+#if DEBUG
+    #Preview {
+        let coordinator = TTSEmptyCoordinator(choices: Array(TTSEmptyCoordinator.kDefaultChoices.prefix(1)))
+        let viewModel = TTSViewViewModel(coordinator: coordinator)
 
-    class TTSEmptyCoordinator: TTSGameplayCoordinatorProtocol {
-        var validationEnabled = CurrentValueSubject<Bool?, Never>(nil)
-
-        func validateUserSelection() {
-            log.debug("Choice validated")
-        }
-
-        var uiModel = CurrentValueSubject<TTSUIModel, Never>(TTSUIModel(action: nil, choices: [
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 1", type: .text, size: 240, state: .idle)),
-        ]))
-
-        func processUserSelection(choiceID: String) {
-            log.debug("\(choiceID)")
-        }
+        return TTSView(viewModel: viewModel)
     }
-
-    let coordinator = TTSEmptyCoordinator()
-    let viewModel = TTSViewViewModel(coordinator: coordinator)
-
-    return TTSView(viewModel: viewModel)
-}
+#endif
