+++ Modules/GameEngineKit/Sources/_NewSystem/Views/TTS/TTS/TTSView.swift
@@ -140,35 +140,11 @@ extension l10n {
     }
 }
 
-#Preview {
-    // MARK: - TTSEmptyCoordinator
+#if DEBUG
+    #Preview {
+        let coordinator = TTSEmptyCoordinator()
+        let viewModel = TTSViewViewModel(coordinator: coordinator)
 
-    class TTSEmptyCoordinator: TTSGameplayCoordinatorProtocol {
-        var uiModel = CurrentValueSubject<TTSUIModel, Never>(TTSUIModel(action: nil, choices: [
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 1", type: .text, size: 240, state: .idle)),
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 2", type: .text,
-                                                                                size: 240, state: .idle)),
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 3\nCorrect", type: .text,
-                                                                                size: 240, state: .correct)),
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "}.triangle.fill", type: .sfsymbol,
-                                                                                size: 240, state: .wrong)),
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 5", type: .text, size: 240, state: .idle)),
-            TTSUIChoiceModel(view: TTSCoordinatorFindTheRightAnswers.ChoiceView(value: "Choice 6", type: .text, size: 240, state: .idle)),
-        ]))
-
-        var validationEnabled = CurrentValueSubject<Bool?, Never>(nil)
-
-        func validateUserSelection() {
-            log.debug("Choice validated")
-        }
-
-        func processUserSelection(choiceID: String) {
-            log.debug("\(choiceID)")
-        }
+        return TTSView(viewModel: viewModel)
     }
-
-    let coordinator = TTSEmptyCoordinator()
-    let viewModel = TTSViewViewModel(coordinator: coordinator)
-
-    return TTSView(viewModel: viewModel)
-}
+#endif
